+++ src/services/mod.rs
@@ -1,1 +1,3 @@
-pub mod transactions;
+pub mod about;
+pub mod transactions_details;
+pub mod transactions_list;
